Defined Admin, Premium and Standard Roles

+++ app/controllers/wikis_controller.rb
@@ -1,18 +1,22 @@
 class WikisController < ApplicationController
   def index
-    @wikis = Wiki.most_recently_updated
+    @wikis = policy_scope(Wiki).most_recently_updated
+    authorize @wikis
   end
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @wiki = current_user.wikis.build(wiki_params)
+    authorize @wiki
     if @wiki.save
       flash[:notice] = "Wiki saved succesfully."
       redirect_to @wiki
@@ -24,10 +28,12 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def update
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
     if @wiki.update_attributes(wiki_params)
       flash[:notice] = "Wiki updated succesfully."
       redirect_to @wiki
@@ -39,8 +45,9 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
     if @wiki.destroy
-      flash[:alert] = "Wiki, #{@wiki.title}, deleted succesfully."
+      flash[:alert] = "Wiki, \"#{@wiki.title}\" deleted succesfully."
       redirect_to @wiki
     else
       flash[:error] = "There was an error deleting wiki. Please try again."
